terminati test

+++ app/Http/Controllers/Backend/ProductController.php
@@ -16,55 +16,82 @@ use Illuminate\Support\Facades\Validator;
 class ProductController extends Controller
 {
     public function pizza(){
-      
         
+        
+        $categories = Category::all();
         $images = Image::all();
         if(empty(session('searchProduct')) && empty(session('category'))){
             $products = Product::paginate(4);
+        }
+        if(session('category_id')){
+            $id = session('category_id');
+            $products = Product::all()->where('category_id' , $id)->paginate(4);
+            
+            session()->forget('category_id');
+            
         }else{
             $q = Product::query();
             $search = session('searchProduct');
             $category = session('category');
-
-            if($search){
-                $q = $q->where('name','LIKE','%'.$search.'%');
-            }
-    
-            if($category){
+            if(($search) && ($category)){
                 foreach($category as $cat){
                  
                     if($cat === "Tutte"){
-                        $q = $q->where('category_id' , '<>' , 'Tutte');
-                    } else{
+                        $q = $q->where('name','LIKE','%'.$search.'%')
+                        ->where('category_id' , '<>' , 'Tutte');
+                    }else{
+                        $q = $q->where('name','LIKE','%'.$search.'%')
+                            ->where('category_id', $cat)
+                            ->where('category_id', $cat);
+                        }
+                    }  
+                }
+                
+                if(($search) && (!$category)){
+                    $q = $q->where('name','LIKE','%'.$search.'%');
                     
-                    $q = $q->where('name','LIKE','%'.$search.'%')
-                        ->where('category_id', $cat)
-                        ->orWhere('category_id', $cat);
-                    }
-                }  
-            }
-        
-            $q = $q->paginate(4);
+                }
+                
+                if((!$search) && ($category)){
+                    //dd($category);
+                
+                    foreach($category as $cat){
+                    
+                        if($cat === "Tutte"){
+                            $q = $q->where('category_id' , '<>' , 'Tutte');
+                        } else{
+                        
+                            $q = $q->where('category_id', $cat)
+                                ->orWhere('category_id', $cat);
+                        }
+                    }  
+                }
+                
+                
+                $q = $q->paginate(4);
+                $products = $q;
+               //dd($products);
+            }    
             
-            $products = $q;
-           
-        }
+       
         
-        $categories = Category::all();
-    
         return view("pizza" , compact('products'))->with(compact('categories'))->with(compact('images'));
     }
 
     public function newProduct(){
-       
-        if (Gate::denies('Gestore')) {
-            return view("welcome")->with("message" , "Non sei Autorizzato ad aggiungere prodotti!");
-        } 
         $categories = Category::all();
+        if (Gate::denies('Gestore')) {
+            abort(403);
+            //return view("welcome" , compact("categories"))->with("message" , "Non sei Autorizzato ad aggiungere prodotti!");
+            
+        } 
         return view("newProduct", compact("categories"));
     }
 
     public function submitProduct(Request $request){
+        if (Gate::denies('Gestore')) {
+            abort(403);            
+        } 
         
         $categories = Category::all();
         $product = new Product();
@@ -82,14 +109,17 @@ class ProductController extends Controller
     }
 
     public function formModify(Product $product){
-        if (Gate::denies('Gestore')) {
-            return view("welcome")->with("message" , "Non sei Autorizzato ad aggiungere prodotti!");
-        } 
         $categories = Category::all();
+        if (Gate::denies('Gestore')) {
+            return view("welcome" , compact("categories"))->with("message" , "Non sei Autorizzato ad aggiungere prodotti!");
+        } 
         return view('modificaProdotto', compact('product'))->with(compact('categories'));
     }
 
     public function modifyProduct(Request $request , Product $product){
+        if (Gate::denies('Gestore')) {
+            abort(403);            
+        } 
         $product->name = $request->name;
         $product->description = $request->description;
         $product->price = $request->price;
@@ -101,19 +131,20 @@ class ProductController extends Controller
     }
 
     public function deleteProduct(Product $product){
-        
+        if (Gate::denies('Gestore')) {
+            abort(403);            
+        } 
         $product->delete();
     
         return redirect(route('pizza'));
     }
 
     public function search(Request $request){
-
         $search = $request->search;
         $category = $request->category_id;
+        // dd($category);
         session()->put('searchProduct' , $search);
         session()->put('category' , $category);
-
         return redirect(route('pizza'));
     }
 
@@ -141,10 +172,16 @@ class ProductController extends Controller
     }
        
     public function addCategory(){
+        if (Gate::denies('Gestore')) {
+            abort(403);
+        } 
         return view('addCategory');
     }
 
     public function addNewCategory(Request $request){
+        if (Gate::denies('Gestore')) {
+            abort(403);
+        } 
         $category = new Category();
         $category->name = $request->input('name');
         $category->img = $request->file('img')->store('public/img');
@@ -156,10 +193,16 @@ class ProductController extends Controller
     }
 
     public function modifyCategory(Category $category){
+        if (Gate::denies('Gestore')) {
+            abort(403);
+        } 
         return view('modifyCategory', compact('category'));
     }
     
     public function modifyNewCategory(Category $category , Request $request){
+        if (Gate::denies('Gestore')) {
+            abort(403);
+        } 
         $category->name = $request->name;
         $category->description = $request->description;
         $category->img = $request->file('img')->store('public/img');
@@ -169,6 +212,9 @@ class ProductController extends Controller
     }
 
     public function deleteCategory(Category $category){
+        if (Gate::denies('Gestore')) {
+            abort(403);
+        } 
         $category->delete();
 
         return redirect(route('welcome'));
